# Conflicts:
#	Build/CMakeLists.txt
#	Include/Rocket/Core/Context.h
#	Include/Rocket/Core/Property.h
#	Include/Rocket/Core/StyleSheetKeywords.h
#	Source/Core/Context.cpp
#	Source/Core/ElementStyle.cpp
#	Source/Core/PropertyDefinition.cpp
#	Source/Core/PropertyParserNumber.cpp
#	Source/Core/StringCache.cpp
#	Source/Core/StringCache.h
#	Source/Core/StyleSheetSpecification.cpp

+++ Source/Core/TransformState.cpp
@@ -34,7 +34,7 @@ namespace Core {
 
 Matrix4f TransformState::Perspective::GetProjection() const throw()
 {
-	float depth = Math::Max(view_size.x, view_size.y);
+	float depth = (float)Math::Max(view_size.x, view_size.y);
 
 	if (distance == 0)
 	{
